@@ -15,53 +15,58 @@
   },
   "private": true,
   "dependencies": {
-    "randomatic": "3.0.0",
-    "lodash": "4.17.10",
-    "@angular/animations": "^5.2.0",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
+    "@angular/animations": "^6.1.8",
+    "@angular/common": "^6.1.8",
+    "@angular/compiler": "^6.1.8",
+    "@angular/core": "^6.1.8",
+    "@angular/forms": "^6.1.8",
+    "@angular/http": "^6.1.8",
+    "@angular/platform-browser": "^6.1.8",
+    "@angular/platform-browser-dynamic": "^6.1.8",
+    "@angular/platform-server": "^6.1.8",
+    "@angular/router": "^6.1.8",
+    "@kyma-project/luigi-client": "0.1.0",
+    "@kyma-project/y-generic-list": "^0.5.2",
     "ace-builds": "^1.3.0",
     "brace": "^0.11.1",
     "copy-webpack-plugin": "^4.3.0",
-    "core-js": "^2.4.1",
+    "core-js": "^2.5.7",
     "js-sha256": "^0.9.0",
+    "lodash": "4.17.10",
     "ng-click-outside": "^3.2.0",
-    "@kyma-project/luigi-client": "0.1.0",
-    "@kyma-project/y-generic-list": "^0.4.10",
     "ng2-ace-editor": "0.3.7",
-    "rxjs": "^5.5.6",
+    "randomatic": "3.0.0",
+    "rxjs": "^6.3.2",
+    "rxjs-compat": "^6.3.2",
     "time-ago-pipe": "^1.2.1",
     "ts-clipboard": "^1.0.16",
     "tslint-angular": "^1.1.1",
     "tslint-config-prettier": "^1.12.0",
-    "zone.js": "^0.8.19"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.6",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
+    "@angular-devkit/build-angular": "^0.8.3",
+    "@angular/cli": "^6.2.3",
+    "@angular/compiler-cli": "^6.1.8",
+    "@angular/language-service": "^6.1.8",
+    "@types/jasmine": "^2.8.8",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
+    "@types/node": "^10.10.1",
     "ajv": "^6.4.0",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "codelyzer": "^4.4.4",
+    "jasmine-core": "^3.2.1",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^3.0.0",
     "karma-chrome-launcher": "^2.2.0",
+    "karma-cli": "^1.0.1",
     "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^1.3.1",
     "karma-phantomjs-launcher": "^1.0.4",
-    "protractor": "^5.4.0",
+    "protractor": "^5.4.1",
+    "rxjs-tslint": "^0.1.5",
     "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "tslint": "^5.11.0",
+    "typescript": "^2.7.2"
   }
 }
