@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 import { GraphqlClientService } from '../graphql-client/graphql-client.service';
 import { AppConfig } from '../app.config';
 import { EventActivation } from '../shared/datamodel/k8s/event-activation';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { Event } from '../shared/datamodel/event';
 import { EventActivationResponse } from '../shared/datamodel/k8s/event-activation-response';
 
