@@ -1,6 +1,7 @@
 /* tslint:disable:no-submodule-imports */
 
-import { Observable } from 'rxjs/Observable';
+import { catchError, timeout } from 'rxjs/operators';
+import { of as observableOf, Observable, forkJoin } from 'rxjs';
 import {
   Component,
   ViewChild,
@@ -31,7 +32,6 @@ import { FetchTokenModalComponent } from '../../fetch-token-modal/fetch-token-mo
 import { ServiceBindingUsagesService } from '../../service-binding-usages/service-binding-usages.service';
 import { ServiceBindingsService } from '../../service-bindings/service-bindings.service';
 import { InstanceBindingState } from '../../shared/datamodel/instance-binding-state';
-import { forkJoin } from 'rxjs/observable/forkJoin';
 import { HttpErrorResponse } from '@angular/common/http';
 import { EventTrigger } from '../../shared/datamodel/event-trigger';
 import { EventActivationsService } from '../../event-activations/event-activations.service';
@@ -43,9 +43,9 @@ import * as randomatic from 'randomatic';
 import * as luigiClient from '@kyma-project/luigi-client';
 
 import { Service } from '../../shared/datamodel/k8s/api-service';
-import { timeout } from 'rxjs/operators';
 import { EventTriggerChooserComponent } from './event-trigger-chooser/event-trigger-chooser.component';
 import { HttpTriggerComponent } from './http-trigger/http-trigger.component';
+
 @Component({
   selector: 'app-lambda-details',
   templateUrl: './lambda-details.component.html',
@@ -98,12 +98,12 @@ export class LambdaDetailsComponent implements AfterViewInit {
   lambda = new Lambda({
     metadata: this.md,
   });
-  loaded: Observable<boolean> = Observable.of(false);
+  loaded: Observable<boolean> = observableOf(false);
   newLabel;
   wrongLabel = false;
   wrongLabelMessage = '';
   error: string = null;
-  hasDependencies: Observable<boolean> = Observable.of(false);
+  hasDependencies: Observable<boolean> = observableOf(false);
   envVarKey = '';
   envVarValue = '';
   isEnvVariableNameInvalid = false;
@@ -184,7 +184,7 @@ export class LambdaDetailsComponent implements AfterViewInit {
             this.title = 'Create Lambda Function';
             this.lambda = this.lambdaDetailsService.initializeLambda();
             this.lambda.spec.function = this.code;
-            this.loaded = Observable.of(true);
+            this.loaded = observableOf(true);
             if (!this.lambda.metadata.name || this.isFunctionNameInvalid) {
               this.editor.setReadOnly(true);
             }
@@ -380,9 +380,11 @@ export class LambdaDetailsComponent implements AfterViewInit {
         createRequests.push(
           this.serviceBindingsService
             .createServiceBinding(serviceBinding, this.token)
-            .catch(err => {
-              return Observable.of(err);
-            }),
+            .pipe(
+              catchError(err => {
+                return observableOf(err);
+              }),
+            ),
         );
       } else {
         serviceBindingUsage.spec.serviceBindingRef.name =
@@ -397,9 +399,11 @@ export class LambdaDetailsComponent implements AfterViewInit {
       createRequests.push(
         this.serviceBindingUsagesService
           .createServiceBindingUsage(serviceBindingUsage, this.token)
-          .catch(err => {
-            return Observable.of(err);
-          }),
+          .pipe(
+            catchError(err => {
+              return observableOf(err);
+            }),
+          ),
       );
     });
 
@@ -420,9 +424,11 @@ export class LambdaDetailsComponent implements AfterViewInit {
                     this.environment,
                     this.token,
                   )
-                  .catch(err => {
-                    return Observable.of(err);
-                  }),
+                  .pipe(
+                    catchError(err => {
+                      return observableOf(err);
+                    }),
+                  ),
               );
             }
           });
@@ -512,9 +518,11 @@ export class LambdaDetailsComponent implements AfterViewInit {
           sub.spec.source_id = trigger.sourceId;
           const req = this.subscriptionsService
             .createSubscription(sub, this.token)
-            .catch(err => {
-              return Observable.of(err);
-            });
+            .pipe(
+              catchError(err => {
+                return observableOf(err);
+              }),
+            );
           createSubscriptionRequests.push(req);
         }
       }
@@ -528,9 +536,11 @@ export class LambdaDetailsComponent implements AfterViewInit {
           this.environment,
           this.token,
         )
-        .catch(err => {
-          return Observable.of(err);
-        });
+        .pipe(
+          catchError(err => {
+            return observableOf(err);
+          }),
+        );
       deleteSubscriptionRequests.push(req);
     });
     this.executeCreateAndDeleteSubscriptionRequests(
@@ -704,12 +714,12 @@ export class LambdaDetailsComponent implements AfterViewInit {
           this.code = lambda.spec.function;
           this.kind = lambda.spec.runtime;
           this.dependency = lambda.spec.deps;
-          this.hasDependencies = Observable.of(
+          this.hasDependencies = observableOf(
             this.dependency != null &&
               this.dependency !== undefined &&
               this.dependency !== '',
           );
-          this.loaded = Observable.of(true);
+          this.loaded = observableOf(true);
         },
         err => {
           this.navigateToList();
@@ -748,7 +758,7 @@ export class LambdaDetailsComponent implements AfterViewInit {
       .getEventActivations(this.environment, this.token)
       .subscribe(
         events => {
-          this.loaded = Observable.of(true);
+          this.loaded = observableOf(true);
         },
         err => {
           this.error = err.message;
@@ -870,12 +880,12 @@ export class LambdaDetailsComponent implements AfterViewInit {
   }
 
   addDependencies() {
-    this.hasDependencies = Observable.of(true);
+    this.hasDependencies = observableOf(true);
   }
   removeDependencies() {
     this.dependency = '';
     this.lambda.spec.deps = null;
-    this.hasDependencies = Observable.of(false);
+    this.hasDependencies = observableOf(false);
   }
 
   /** validatesName checks whether a function name is abiding by RFC 1123 or not */
