@@ -1,5 +1,5 @@
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/catch';
+import { refCount, publishReplay, map, catchError } from 'rxjs/operators';
+import { of as observableOf, Observable, forkJoin } from 'rxjs';
 
 import { Component, ViewChild, ChangeDetectorRef } from '@angular/core';
 import { IFunction, Lambda } from './../../shared/datamodel/k8s/function';
@@ -22,7 +22,6 @@ import { DeploymentDetailsService } from './deployment-details.service';
 
 import { AppConfig } from '../../app.config';
 import { ApisService } from '../../apis/apis.service';
-import { forkJoin } from 'rxjs/observable/forkJoin';
 import { ServiceBindingUsagesService } from '../../service-binding-usages/service-binding-usages.service';
 import { SubscriptionsService } from '../../subscriptions/subscriptions.service';
 import { Subscription } from '../../shared/datamodel/k8s/subscription';
@@ -138,9 +137,11 @@ export class LambdasComponent extends GenericTableComponent {
                 this.environment,
                 this.token,
               )
-              .catch(err => {
-                return Observable.of(err);
-              }),
+              .pipe(
+                catchError(err => {
+                  return observableOf(err);
+                }),
+              ),
           );
         });
 
@@ -171,9 +172,11 @@ export class LambdasComponent extends GenericTableComponent {
                 this.environment,
                 this.token,
               )
-              .catch(err => {
-                return Observable.of(err);
-              }),
+              .pipe(
+                catchError(err => {
+                  return observableOf(err);
+                }),
+              ),
           );
         });
         forkJoin(deleteRequests).subscribe(responses => {
@@ -194,14 +197,18 @@ export class LambdasComponent extends GenericTableComponent {
         const func = new Lambda(entry);
         func.functionStatus = that.deploymentDetailsService
           .getDeployments(func.getName(), that.environment, that.token)
-          .map(deploymentList => {
-            return deploymentList.items[0].status;
-          })
-          .catch(() => {
-            return Observable.of(0);
-          })
-          .publishReplay(1)
-          .refCount();
+          .pipe(
+            map(deploymentList => {
+              return deploymentList.items[0].status;
+            }),
+            catchError(() => {
+              return observableOf(0);
+            }),
+          )
+          .pipe(
+            publishReplay(1),
+            refCount(),
+          );
         return func;
       },
     };
